pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kali-ai-orchestrator"
version = "0.1.0"
description = "AI-powered orchestrator for Kali Linux pentesting tools with hexstrike-ai MCP integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Kali AI Orchestrator Contributors"}
]
keywords = ["pentesting", "security", "ai", "kali", "mcp", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.7.0",
    "textual>=0.65.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "httpx>=0.25.2",
    "google-generativeai>=0.3.2",
    "weasyprint>=60.2",
    "ipaddress>=1.0.23",
    "netaddr>=0.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-mock>=3.12.0",
    "black>=23.12.1",
    "ruff>=0.1.9",
    "mypy>=1.7.1",
]

[project.scripts]
kali-orchestrator = "kali_orchestrator.main:app"

[tool.setuptools]
packages = ["kali_orchestrator", "kali_orchestrator.tools", "kali_orchestrator.plugins", "kali_orchestrator.personas"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

